Fixed broken radius-rect computation Always use the double type for coordinate related data

+++ src/map/mapfile.cpp
@@ -54,13 +54,13 @@ int MapFile::parse(QIODevice &device, QList<CalibrationPoint> &points,
 				int latd = list.at(6).trimmed().toInt(&res);
 				if (!res)
 					ll = false;
-				qreal latm = list.at(7).trimmed().toFloat(&res);
+				double latm = list.at(7).trimmed().toDouble(&res);
 				if (!res)
 					ll = false;
 				int lond = list.at(9).trimmed().toInt(&res);
 				if (!res)
 					ll = false;
-				qreal lonm = list.at(10).trimmed().toFloat(&res);
+				double lonm = list.at(10).trimmed().toDouble(&res);
 				if (!res)
 					ll = false;
 				if (ll && list.at(8).trimmed() == "S") {
@@ -75,10 +75,10 @@ int MapFile::parse(QIODevice &device, QList<CalibrationPoint> &points,
 				p.zone = list.at(13).trimmed().toInt(&res);
 				if (!res)
 					p.zone = 0;
-				qreal ppx = list.at(14).trimmed().toFloat(&res);
+				double ppx = list.at(14).trimmed().toDouble(&res);
 				if (!res)
 					pp = false;
-				qreal ppy = list.at(15).trimmed().toFloat(&res);
+				double ppy = list.at(15).trimmed().toDouble(&res);
 				if (!res)
 					pp = false;
 				if (list.at(16).trimmed() == "S")
